pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pka-dashboard"
version = "0.1.0"
description = "Renarin — read-only web dashboard for Scott's Personal Knowledge Assistant"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
    "mistune>=3.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.0",
    "anyio>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
